ticketing-system/backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
anthropic==0.23.1
python-dotenv==1.0.1
pydantic==2.6.1
websockets==12.0
python-multipart==0.0.9
apscheduler==3.10.4
